refacor the repo
