+++ dflash/benchmark.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import argparse
+import json
 import os
 import random
 import re
@@ -9,160 +10,162 @@ import time
 import warnings
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from itertools import chain
+from pathlib import Path
 from types import SimpleNamespace
-from typing import Any, List, Optional
+from typing import Any
 
 import numpy as np
 import requests
-import torch
-from torch import distributed as torch_dist
 from loguru import logger
 from rich import print
 from tqdm import tqdm
-from transformers import AutoModelForCausalLM, AutoTokenizer, DynamicCache
 
-from .model import DFlashDraftModel, sample, load_and_process_dataset, extract_context_feature
+random.seed(42)
 
 
-def _dist_init() -> None:
+CACHE_DIR = Path(__file__).parent.parent / "cache"
+
+DATASETS = {
+    "gsm8k": {
+        "load_args": ("openai/gsm8k", "main"),
+        "load_kwargs": {"split": "test"},
+        "format": lambda x: "{question}\nPlease reason step by step, and put your final answer within \\boxed{{}}.".format(**x),
+    },
+    "math500": {
+        "load_args": ("HuggingFaceH4/MATH-500",),
+        "load_kwargs": {"split": "test"},
+        "format": lambda x: "{problem}\nPlease reason step by step, and put your final answer within \\boxed{{}}.".format(**x),
+    },
+    "humaneval": {
+        "load_args": ("openai/openai_humaneval",),
+        "load_kwargs": {"split": "test"},
+        "format": lambda x: "Write a solution to the following problem and make sure that it passes the tests:\n```python\n{prompt}\n```".format(**x),
+    },
+    "mbpp": {
+        "load_args": ("google-research-datasets/mbpp", "sanitized"),
+        "load_kwargs": {"split": "test"},
+        "format": lambda x: x["prompt"],
+    },
+    "mt-bench": {
+        "load_args": ("HuggingFaceH4/mt_bench_prompts",),
+        "load_kwargs": {"split": "train"},
+        "format": lambda x: x["prompt"],
+        "multi_turn": True,
+    },
+}
+
+
+def _prepare_dataset(name: str) -> Path:
+    from datasets import load_dataset
+
+    cfg = DATASETS[name]
+    CACHE_DIR.mkdir(exist_ok=True)
+    out_path = CACHE_DIR / f"{name}.jsonl"
+
+    print(f"[download] {name} ...")
+    dataset = load_dataset(*cfg["load_args"], **cfg["load_kwargs"])
+
+    with open(out_path, "w") as f:
+        for row in dataset:
+            if cfg.get("multi_turn"):
+                turns = cfg["format"](row)
+            else:
+                turns = [cfg["format"](row)]
+            f.write(json.dumps({"turns": turns}) + "\n")
+
+    with open(out_path) as f:
+        num_samples = sum(1 for _ in f)
+    print(f"[cached] {out_path}  ({num_samples} samples)")
+    return out_path
+
+
+def load_and_process_dataset(data_name: str) -> list[dict]:
+    if data_name not in DATASETS:
+        raise ValueError(f"Unknown dataset '{data_name}'. Available: {list(DATASETS.keys())}")
+
+    path = CACHE_DIR / f"{data_name}.jsonl"
+    if not path.exists():
+        _prepare_dataset(data_name)
+
+    with open(path) as f:
+        return [json.loads(line) for line in f]
+
+
+def _limit_dataset(dataset: list[dict], max_samples: int | None) -> list[dict]:
+    if max_samples is None or len(dataset) <= max_samples:
+        return dataset
+    random.shuffle(dataset)
+    return dataset[:max_samples]
+
+
+def _apply_chat_template(tokenizer, messages: list[dict], enable_thinking: bool) -> str:
+    return tokenizer.apply_chat_template(
+        messages,
+        tokenize=False,
+        add_generation_prompt=True,
+        enable_thinking=enable_thinking,
+    )
+
+
+def _make_decode_metrics(num_output_tokens: int, generation_tps: float, acceptance_lengths: list[int]) -> SimpleNamespace:
+    return SimpleNamespace(
+        num_output_tokens=num_output_tokens,
+        time_per_output_token=1.0 / generation_tps if generation_tps > 0 else float("inf"),
+        acceptance_lengths=acceptance_lengths,
+    )
+
+
+def _print_decode_summary(responses: list[dict[int, SimpleNamespace]], block_size: int) -> None:
+    baseline_tpot = np.mean([r[1].time_per_output_token for r in responses])
+    dflash_tpot = np.mean([r[block_size].time_per_output_token for r in responses])
+    print(f"Baseline throughput: {1 / baseline_tpot:.2f} tok/s")
+    print(f"DFlash throughput:  {1 / dflash_tpot:.2f} tok/s")
+    print(f"Decoding speedup: {baseline_tpot / dflash_tpot:.2f}")
+
+    mean_accept = np.mean([np.mean(r[block_size].acceptance_lengths) for r in responses])
+    print(f"Average Acceptance length: {mean_accept:.2f}")
+
+    acceptance_lengths = list(chain.from_iterable(r[block_size].acceptance_lengths for r in responses))
+    histogram = [acceptance_lengths.count(b) / len(acceptance_lengths) for b in range(block_size + 1)]
+    print(f"Acceptance length histogram: {[f'{x * 100:.1f}%' for x in histogram]}")
+
+
+def _env_int(name: str, default: int) -> int:
+    return int(os.environ.get(name, default))
+
+
+def _dist_init(torch_dist) -> None:
     if "RANK" not in os.environ:
         warnings.warn("RANK not set. Skipping distributed initialization.")
         return
     torch_dist.init_process_group(backend="nccl", init_method="env://")
 
+
 def _dist_size() -> int:
-    return int(os.environ.get("WORLD_SIZE", 1))
+    return _env_int("WORLD_SIZE", 1)
+
 
 def _dist_rank() -> int:
-    return int(os.environ.get("RANK", 0))
+    return _env_int("RANK", 0)
+
 
 def _dist_local_rank() -> int:
-    return int(os.environ.get("LOCAL_RANK", 0))
+    return _env_int("LOCAL_RANK", 0)
+
 
 def _dist_is_main() -> bool:
     return _dist_rank() == 0
 
-def _dist_gather(obj: Any, dst: int = 0) -> Optional[List[Any]]:
+
+def _dist_gather(torch_dist, obj: Any, dst: int = 0):
     if not torch_dist.is_initialized():
         return [obj]
     if _dist_is_main():
-        objs: List[Any] = [None for _ in range(_dist_size())]
+        objs = [None for _ in range(_dist_size())]
         torch_dist.gather_object(obj, objs, dst=dst)
         return objs
-    else:
-        torch_dist.gather_object(obj, dst=dst)
-        return None
-
-
-def _cuda_time() -> float:
-    torch.cuda.synchronize()
-    return time.perf_counter()
-
-
-@torch.inference_mode()
-def _dflash_generate(
-    model: DFlashDraftModel,
-    target: AutoModelForCausalLM,
-    input_ids: torch.Tensor,
-    mask_token_id: int,
-    max_new_tokens: int,
-    block_size: int,
-    stop_token_ids: list[int],
-    temperature: float = 0.0,
-) -> SimpleNamespace:
-    num_input_tokens = input_ids.shape[1]
-    max_length = num_input_tokens + max_new_tokens
-
-    output_ids = torch.full(
-        (1, max_length + block_size), mask_token_id, dtype=torch.long, device=model.device,
-    )
-    position_ids = torch.arange(output_ids.shape[1], device=model.device).unsqueeze(0)
-    past_key_values_target = DynamicCache()
-    past_key_values_draft = DynamicCache()
-
-    prefill_start = _cuda_time()
-    output = target(
-        input_ids,
-        position_ids=position_ids[:, :num_input_tokens],
-        past_key_values=past_key_values_target,
-        use_cache=True,
-        logits_to_keep=1,
-        output_hidden_states=True if block_size > 1 else False,
-    )
-    output_ids[:, :num_input_tokens] = input_ids
-    output_ids[:, num_input_tokens:num_input_tokens+1] = sample(output.logits, temperature)
-    if block_size > 1:
-        target_hidden = extract_context_feature(output.hidden_states, model.target_layer_ids)
-    time_to_first_token = _cuda_time() - prefill_start
-
-    decode_start = _cuda_time()
-    start = input_ids.shape[1]
-    acceptance_lengths = []
-    draft_prefill = True
-
-    while start < max_length:
-        block_output_ids = output_ids[:, start : start + block_size].clone()
-        block_position_ids = position_ids[:, start : start + block_size]
-        if block_size > 1:
-            noise_embedding = target.model.embed_tokens(block_output_ids)
-            draft_logits = target.lm_head(model(
-                target_hidden=target_hidden,
-                noise_embedding=noise_embedding,
-                position_ids=position_ids[:, past_key_values_draft.get_seq_length(): start + block_size],
-                past_key_values=past_key_values_draft,
-                use_cache=True,
-                is_causal=False,
-            )[:, -block_size+1:, :])
-            past_key_values_draft.crop(start)
-            block_output_ids[:, 1:] = sample(draft_logits)
-            if draft_prefill:
-                draft_prefill = False
-                decode_start = _cuda_time()
-
-        output = target(
-            block_output_ids,
-            position_ids=block_position_ids,
-            past_key_values=past_key_values_target,
-            use_cache=True,
-            output_hidden_states=True if block_size > 1 else False,
-        )
-
-        posterior = sample(output.logits, temperature)
-        acceptance_length = (block_output_ids[:, 1:] == posterior[:, :-1]).cumprod(dim=1).sum(dim=1)[0].item()
-        output_ids[:, start : start + acceptance_length + 1] = block_output_ids[:, : acceptance_length + 1]
-        output_ids[:, start + acceptance_length + 1] = posterior[:, acceptance_length]
-
-        acceptance_lengths.append(acceptance_length + 1)
-        start += acceptance_length + 1
-        past_key_values_target.crop(start)
-        if block_size > 1:
-            target_hidden = extract_context_feature(output.hidden_states, model.target_layer_ids)[:, :acceptance_length + 1, :]
-
-        if stop_token_ids is not None and any(
-            stop_token_id in output_ids[:, num_input_tokens:] for stop_token_id in stop_token_ids
-        ):
-            break
-
-    output_ids = output_ids[:, :max_length]
-    output_ids = output_ids[:, output_ids[0] != mask_token_id]
-    if stop_token_ids is not None:
-        stop_token_ids = torch.tensor(stop_token_ids, device=output_ids.device)
-        stop_token_indices = torch.isin(output_ids[0][num_input_tokens:], stop_token_ids).nonzero(as_tuple=True)[0]
-        if stop_token_indices.numel() > 0:
-            output_ids = output_ids[:, : num_input_tokens + stop_token_indices[0] + 1]
-
-    num_output_tokens = output_ids.shape[1] - num_input_tokens
-    total_decode_time = _cuda_time() - decode_start
-    time_per_output_token = total_decode_time / num_output_tokens
-
-    return SimpleNamespace(
-        output_ids=output_ids,
-        num_input_tokens=num_input_tokens,
-        num_output_tokens=num_output_tokens,
-        time_to_first_token=time_to_first_token,
-        time_per_output_token=time_per_output_token,
-        acceptance_lengths=acceptance_lengths,
-    )
+    torch_dist.gather_object(obj, dst=dst)
+    return None
 
 
 _TRANSFORMERS_SUPPORTED_PATTERN = re.compile(r"qwen3(?!\.5)[\w-]*|llama.*3\.1.*8b.*instruct", re.IGNORECASE)
@@ -177,7 +180,26 @@ def _check_transformers_model(model_name: str) -> None:
         )
 
 
+def _get_transformers_attn_impl() -> str:
+    try:
+        import flash_attn  # noqa: F401
+        return "flash_attention_2"
+    except ImportError:
+        logger.warning(
+            "flash_attn not installed. Falling back to torch.sdpa. Speedup will be lower. "
+            "For optimal speedup in Transformers backend, please install: "
+            "pip install flash-attn --no-build-isolation"
+        )
+        return "sdpa"
+
+
 def _run_transformers(args: argparse.Namespace) -> None:
+    import torch
+    from torch import distributed as torch_dist
+    from transformers import AutoModelForCausalLM, AutoTokenizer
+
+    from .model import DFlashDraftModel, dflash_generate
+
     _check_transformers_model(args.model)
 
     random.seed(0)
@@ -187,24 +209,10 @@ def _run_transformers(args: argparse.Namespace) -> None:
     torch.backends.cudnn.deterministic = True
     torch.backends.cudnn.benchmark = False
 
-    _dist_init()
+    _dist_init(torch_dist)
     torch.cuda.set_device(_dist_local_rank())
     device = torch.device(f"cuda:{_dist_local_rank()}")
-
-    def has_flash_attn():
-        try:
-            import flash_attn
-            return True
-        except ImportError:
-            logger.warning(
-                "flash_attn not installed. Falling back to torch.sdpa. Speedup will be lower. "
-                "For optimal speedup in Transformers backend, please install: "
-                "pip install flash-attn --no-build-isolation"
-            )
-            return False
-
-    installed_flash_attn = has_flash_attn()
-    attn_impl = "flash_attention_2" if installed_flash_attn else "sdpa"
+    attn_impl = _get_transformers_attn_impl()
 
     target = AutoModelForCausalLM.from_pretrained(
         args.model, attn_implementation=attn_impl, dtype=torch.bfloat16,
@@ -218,9 +226,7 @@ def _run_transformers(args: argparse.Namespace) -> None:
     tokenizer = AutoTokenizer.from_pretrained(args.model)
     dataset = load_and_process_dataset(args.dataset)
 
-    if args.max_samples is not None and len(dataset) > args.max_samples:
-        random.shuffle(dataset)
-        dataset = dataset[:args.max_samples]
+    dataset = _limit_dataset(dataset, args.max_samples)
 
     responses = []
     indices = range(_dist_rank(), len(dataset), _dist_size())
@@ -229,18 +235,20 @@ def _run_transformers(args: argparse.Namespace) -> None:
         messages = []
         for user_content in instance["turns"]:
             messages.append({"role": "user", "content": user_content})
-            input_text = tokenizer.apply_chat_template(
-                messages, tokenize=False, add_generation_prompt=True, enable_thinking=False,
-            )
+            input_text = _apply_chat_template(tokenizer, messages, args.enable_thinking)
             input_ids = tokenizer.encode(input_text, return_tensors="pt").to(target.device)
 
             response = {}
             for bs in [1, block_size]:
-                response[bs] = _dflash_generate(
-                    model=draft_model, target=target, input_ids=input_ids,
-                    mask_token_id=draft_model.mask_token_id,
-                    max_new_tokens=args.max_new_tokens, block_size=bs,
-                    stop_token_ids=[tokenizer.eos_token_id], temperature=args.temperature,
+                response[bs] = dflash_generate(
+                    draft_model,
+                    target=target,
+                    input_ids=input_ids,
+                    max_new_tokens=args.max_new_tokens,
+                    stop_token_ids=[tokenizer.eos_token_id],
+                    temperature=args.temperature,
+                    block_size=bs,
+                    return_stats=True,
                 )
 
             spec_response = response[block_size]
@@ -250,34 +258,33 @@ def _run_transformers(args: argparse.Namespace) -> None:
             responses.append(response)
 
     if _dist_size() > 1:
-        responses = _dist_gather(responses, dst=0)
+        responses = _dist_gather(torch_dist, responses, dst=0)
         if not _dist_is_main():
             return
         responses = list(chain(*responses))
 
-    t1 = np.mean([r[1].time_per_output_token for r in responses])
-    tb = np.mean([r[block_size].time_per_output_token for r in responses])
-    print(f"Decoding speedup: {t1 / tb:.2f}")
-
-    tau = np.mean([np.mean(r[block_size].acceptance_lengths) for r in responses])
-    print(f"Average Acceptance length: {tau:.2f}")
-
-    acceptance_lengths = list(chain(*[r[block_size].acceptance_lengths for r in responses]))
-    histogram = [acceptance_lengths.count(b) / len(acceptance_lengths) for b in range(block_size + 1)]
-    print(f"Acceptance length histogram: {[f'{x * 100:.1f}%' for x in histogram]}")
+    _print_decode_summary(responses, block_size)
 
 
 def _send_sglang(
-    base_url: str, text: str, *, max_new_tokens: int,
-    temperature: float, top_p: float, top_k: int, timeout_s: int,
+    base_url: str,
+    text: str,
+    *,
+    max_new_tokens: int,
+    temperature: float,
+    top_p: float,
+    top_k: int,
+    timeout_s: int,
 ) -> dict:
     resp = requests.post(
         base_url + "/generate",
         json={
             "text": text,
             "sampling_params": {
-                "temperature": temperature, "top_p": top_p,
-                "top_k": top_k, "max_new_tokens": max_new_tokens,
+                "temperature": temperature,
+                "top_p": top_p,
+                "top_k": top_k,
+                "max_new_tokens": max_new_tokens,
             },
         },
         timeout=timeout_s,
@@ -288,8 +295,15 @@ def _send_sglang(
 
 
 def _send_vllm(
-    base_url: str, text: str, *, model: str, max_new_tokens: int,
-    temperature: float, top_p: float, timeout_s: int,
+    base_url: str,
+    text: str,
+    *,
+    model: str,
+    max_new_tokens: int,
+    temperature: float,
+    top_p: float,
+    top_k: int,
+    timeout_s: int,
     enable_thinking: bool = False,
 ) -> dict:
     body: dict = {
@@ -298,6 +312,7 @@ def _send_vllm(
         "max_tokens": max_new_tokens,
         "temperature": temperature,
         "top_p": top_p,
+        "top_k": top_k,
         "chat_template_kwargs": {"enable_thinking": enable_thinking},
     }
     resp = requests.post(
@@ -310,10 +325,10 @@ def _send_vllm(
 
 
 def _run_mlx(args: argparse.Namespace) -> None:
-    from .utils_mlx import load, load_draft
-    from .generate_mlx import stream_generate
     from mlx_lm import stream_generate as stream_generate_baseline
     from mlx_lm.sample_utils import make_sampler
+
+    from .model_mlx import load, load_draft, stream_generate
 
     sampler = make_sampler(temp=args.temperature)
 
@@ -324,11 +339,8 @@ def _run_mlx(args: argparse.Namespace) -> None:
     block_size = args.block_size if args.block_size is not None else int(draft.config.block_size)
 
     dataset = load_and_process_dataset(args.dataset)
-    if args.max_samples is not None and len(dataset) > args.max_samples:
-        random.shuffle(dataset)
-        dataset = dataset[:args.max_samples]
+    dataset = _limit_dataset(dataset, args.max_samples)
 
-    # Warmup
     list(stream_generate_baseline(model, tokenizer, tokenizer.encode("Hi"), 3))
 
     responses = []
@@ -337,55 +349,38 @@ def _run_mlx(args: argparse.Namespace) -> None:
         messages = []
         for user_content in instance["turns"]:
             messages.append({"role": "user", "content": user_content})
-            prompt = tokenizer.apply_chat_template(
-                messages, tokenize=False, add_generation_prompt=True, enable_thinking=False,
-            )
+            prompt = _apply_chat_template(tokenizer, messages, args.enable_thinking)
 
             response = {}
 
-            # Baseline (block_size=1)
             tokens_bl, tps_bl = [], 0
             for r in stream_generate_baseline(model, tokenizer, prompt, args.max_new_tokens, sampler=sampler):
                 tokens_bl.append(r.token)
                 tps_bl = r.generation_tps
-            response[1] = SimpleNamespace(
-                num_output_tokens=len(tokens_bl),
-                time_per_output_token=1.0 / tps_bl if tps_bl > 0 else float("inf"),
-                acceptance_lengths=[1],
-            )
+            response[1] = _make_decode_metrics(len(tokens_bl), tps_bl, [1])
 
-            # DFlash
             tokens_df, accs, tps_df = [], [], 0
-            for r in stream_generate(model, draft, tokenizer, prompt, block_size, args.max_new_tokens, sampler):
+            for r in stream_generate(model, draft, tokenizer, prompt, block_size, args.max_new_tokens, sampler=sampler):
                 tokens_df.extend(r.tokens)
                 accs.append(r.accepted)
                 tps_df = r.generation_tps
-            response[block_size] = SimpleNamespace(
-                num_output_tokens=len(tokens_df),
-                time_per_output_token=1.0 / tps_df if tps_df > 0 else float("inf"),
-                acceptance_lengths=accs,
-            )
+            response[block_size] = _make_decode_metrics(len(tokens_df), tps_df, accs)
 
             output_text = tokenizer.decode(tokens_df)
             messages.append({"role": "assistant", "content": output_text})
             responses.append(response)
 
-    t1 = np.mean([r[1].time_per_output_token for r in responses])
-    tb = np.mean([r[block_size].time_per_output_token for r in responses])
-    print(f"Decoding speedup: {t1 / tb:.2f}")
-
-    tau = np.mean([np.mean(r[block_size].acceptance_lengths) for r in responses])
-    print(f"Average Acceptance length: {tau:.2f}")
-
-    acceptance_lengths = list(chain(*[r[block_size].acceptance_lengths for r in responses]))
-    histogram = [acceptance_lengths.count(b) / len(acceptance_lengths) for b in range(block_size + 1)]
-    print(f"Acceptance length histogram: {[f'{x * 100:.1f}%' for x in histogram]}")
+    _print_decode_summary(responses, block_size)
 
 
 def _run_server(args: argparse.Namespace) -> None:
     is_vllm = args.backend == "vllm"
     dataset = load_and_process_dataset(args.dataset)
-    tokenizer = AutoTokenizer.from_pretrained(args.model, trust_remote_code=True)
+    tokenizer = None
+    if not is_vllm:
+        from transformers import AutoTokenizer
+
+        tokenizer = AutoTokenizer.from_pretrained(args.model, trust_remote_code=True)
 
     num_prompts = args.num_prompts + args.concurrency
     prompts: list[str] = []
@@ -395,28 +390,34 @@ def _run_server(args: argparse.Namespace) -> None:
         if is_vllm:
             prompts.append(user_content)
         else:
-            prompts.append(tokenizer.apply_chat_template(
+            prompts.append(_apply_chat_template(
+                tokenizer,
                 [{"role": "user", "content": user_content}],
-                tokenize=False, add_generation_prompt=True,
-                enable_thinking=args.enable_thinking,
+                args.enable_thinking,
             ))
 
     def send_one(prompt: str) -> dict:
         if is_vllm:
             return _send_vllm(
-                args.base_url, prompt, model=args.model,
+                args.base_url,
+                prompt,
+                model=args.model,
                 max_new_tokens=args.max_new_tokens,
-                temperature=args.temperature, top_p=args.top_p,
+                temperature=args.temperature,
+                top_p=args.top_p,
+                top_k=args.top_k,
                 timeout_s=args.timeout_s,
                 enable_thinking=args.enable_thinking,
             )
-        else:
-            return _send_sglang(
-                args.base_url, prompt,
-                max_new_tokens=args.max_new_tokens,
-                temperature=args.temperature, top_p=args.top_p,
-                top_k=args.top_k, timeout_s=args.timeout_s,
-            )
+        return _send_sglang(
+            args.base_url,
+            prompt,
+            max_new_tokens=args.max_new_tokens,
+            temperature=args.temperature,
+            top_p=args.top_p,
+            top_k=args.top_k,
+            timeout_s=args.timeout_s,
+        )
 
     if not is_vllm:
         try:
@@ -457,7 +458,7 @@ def _run_server(args: argparse.Namespace) -> None:
     latency = time.perf_counter() - start
     toks_per_s = total_tokens / max(latency, 1e-6)
 
-    print(f"\n{'='*50}")
+    print(f"\n{'=' * 50}")
     print(f"Backend:          {args.backend}")
     print(f"Dataset:          {args.dataset}")
     print(f"Num prompts:      {args.num_prompts}")
@@ -469,7 +470,7 @@ def _run_server(args: argparse.Namespace) -> None:
         print(f"Accept length:    {statistics.mean(spec_accept_lengths):.3f}")
     if spec_verify_ct_sum > 0:
         print(f"Spec verify ct:   {spec_verify_ct_sum}")
-    print(f"{'='*50}")
+    print(f"{'=' * 50}")
 
 
 def main() -> None:
@@ -493,6 +494,11 @@ def main() -> None:
     parser.add_argument("--timeout-s", type=int, default=3600)
 
     args = parser.parse_args()
+
+    assert not (args.enable_thinking and any(x in args.model.lower() for x in ["qwen3-4b", "qwen3-8b"])), (
+        "DFlash draft models for Qwen3-4B and Qwen3-8B were not trained with thinking traces. "
+        "Using --enable-thinking will lead to suboptimal performance."
+    )
 
     if args.backend == "transformers":
         if args.draft_model is None:
