+++ dflash/model_mlx.py
@@ -1,10 +1,23 @@
+import json
+import time
 from dataclasses import dataclass
-from typing import List, Tuple
+from pathlib import Path
+from typing import List, Optional, Tuple
 
 import mlx.core as mx
 import mlx.nn as nn
-from mlx_lm.models.cache import KVCache
+from huggingface_hub import snapshot_download
+from mlx_lm.generate import generation_stream
+from mlx_lm.models.cache import KVCache, can_trim_prompt_cache, make_prompt_cache, trim_prompt_cache
 from mlx_lm.models.qwen3 import MLP
+from mlx_lm.sample_utils import make_sampler
+from mlx_lm.tokenizer_utils import TokenizerWrapper
+
+try:
+    import mlx_lm.models.gated_delta as _gd_mod
+    _HAS_GDN = True
+except ImportError:
+    _HAS_GDN = False
 
 
 @dataclass
@@ -26,47 +39,35 @@ class DFlashConfig:
 
 
 class DFlashAttention(nn.Module):
-    """DFlash cross-attention: Q from noise, K/V from [target_hidden, noise]."""
-
     def __init__(self, config: DFlashConfig):
         super().__init__()
-
         dim = config.hidden_size
         self.n_heads = n_heads = config.num_attention_heads
         self.n_kv_heads = n_kv_heads = config.num_key_value_heads
+        self.scale = config.head_dim ** -0.5
+        self.q_proj = nn.Linear(dim, n_heads * config.head_dim, bias=False)
+        self.k_proj = nn.Linear(dim, n_kv_heads * config.head_dim, bias=False)
+        self.v_proj = nn.Linear(dim, n_kv_heads * config.head_dim, bias=False)
+        self.o_proj = nn.Linear(n_heads * config.head_dim, dim, bias=False)
+        self.q_norm = nn.RMSNorm(config.head_dim, eps=config.rms_norm_eps)
+        self.k_norm = nn.RMSNorm(config.head_dim, eps=config.rms_norm_eps)
 
-        head_dim = config.head_dim
-        self.scale = head_dim**-0.5
-
-        self.q_proj = nn.Linear(dim, n_heads * head_dim, bias=False)
-        self.k_proj = nn.Linear(dim, n_kv_heads * head_dim, bias=False)
-        self.v_proj = nn.Linear(dim, n_kv_heads * head_dim, bias=False)
-        self.o_proj = nn.Linear(n_heads * head_dim, dim, bias=False)
-        self.q_norm = nn.RMSNorm(head_dim, eps=config.rms_norm_eps)
-        self.k_norm = nn.RMSNorm(head_dim, eps=config.rms_norm_eps)
-
-    def __call__(self, x: mx.array, x_ctx: mx.array, rope: nn.RoPE, cache: KVCache) -> mx.array:
+    def __call__(self, x, x_ctx, rope, cache):
         B, L, _ = x.shape
         S = x_ctx.shape[1]
-
         c = mx.concatenate([x_ctx, x], axis=1)
         queries, keys, values = self.q_proj(x), self.k_proj(c), self.v_proj(c)
-
         queries = self.q_norm(queries.reshape(B, L, self.n_heads, -1)).transpose(0, 2, 1, 3)
         keys = self.k_norm(keys.reshape(B, S + L, self.n_kv_heads, -1)).transpose(0, 2, 1, 3)
         values = values.reshape(B, S + L, self.n_kv_heads, -1).transpose(0, 2, 1, 3)
-
         queries = rope(queries, offset=cache.offset + S)
         keys = rope(keys, offset=cache.offset)
         keys, values = cache.update_and_fetch(keys, values)
-
         output = mx.fast.scaled_dot_product_attention(queries, keys, values, scale=self.scale)
         return self.o_proj(output.transpose(0, 2, 1, 3).reshape(B, L, -1))
 
 
 class DFlashDecoderLayer(nn.Module):
-    """Single DFlash transformer layer with pre-norm."""
-
     def __init__(self, config: DFlashConfig):
         super().__init__()
         self.self_attn = DFlashAttention(config)
@@ -74,35 +75,25 @@ class DFlashDecoderLayer(nn.Module):
         self.input_layernorm = nn.RMSNorm(config.hidden_size, eps=config.rms_norm_eps)
         self.post_attention_layernorm = nn.RMSNorm(config.hidden_size, eps=config.rms_norm_eps)
 
-    def __call__(self, x: mx.array, x_ctx: mx.array, rope: nn.RoPE, cache: KVCache) -> mx.array:
-        r = self.self_attn(self.input_layernorm(x), x_ctx, rope, cache)
-        h = x + r
-        r = self.mlp(self.post_attention_layernorm(h))
-        out = h + r
-        return out
+    def __call__(self, x, x_ctx, rope, cache):
+        h = x + self.self_attn(self.input_layernorm(x), x_ctx, rope, cache)
+        return h + self.mlp(self.post_attention_layernorm(h))
 
 
 class DFlashDraftModel(nn.Module):
-    """DFlash draft model: predicts multiple tokens in parallel using target hidden states."""
-
     def __init__(self, config: DFlashConfig):
         super().__init__()
         self.config = config
-
         concat_dim = len(config.target_layer_ids) * config.hidden_size
         self.fc = nn.Linear(concat_dim, config.hidden_size, bias=False)
         self.hidden_norm = nn.RMSNorm(config.hidden_size, eps=config.rms_norm_eps)
-
         self.layers = [DFlashDecoderLayer(config) for _ in range(config.num_hidden_layers)]
         self.norm = nn.RMSNorm(config.hidden_size, eps=config.rms_norm_eps)
         self.rope = nn.RoPE(config.head_dim, traditional=False, base=config.rope_theta)
-
-        # Set by bind()
         self.embed_tokens = None
         self.lm_head = None
 
-    def bind(self, target_model: nn.Module) -> "DFlashDraftModel":
-        """Bind to target model's embedding and lm_head layers."""
+    def bind(self, target_model):
         if hasattr(target_model, "embed_tokens"):
             inner = target_model
         elif hasattr(target_model, "model") and hasattr(target_model.model, "embed_tokens"):
@@ -115,20 +106,280 @@ class DFlashDraftModel(nn.Module):
             raise AttributeError(f"Cannot find embed_tokens in {type(target_model).__name__}")
         self.embed_tokens = inner.embed_tokens
         lm = getattr(target_model, "language_model", target_model)
-        lm_head = getattr(target_model, "lm_head", None) or getattr(lm, "lm_head", None)
-        self.lm_head = lm_head or self.embed_tokens.as_linear
+        self.lm_head = getattr(target_model, "lm_head", None) or getattr(lm, "lm_head", None) or self.embed_tokens.as_linear
         return self
 
-    def make_cache(self) -> List[KVCache]:
-        """Create KV cache for generation."""
+    def make_cache(self):
         return [KVCache() for _ in self.layers]
 
-    def __call__(self, inputs: mx.array, target_hidden: mx.array, cache: List[KVCache]) -> mx.array:
-        """Forward: inputs [B, L] + target_hidden [B, S, hidden*n] -> logits [B, L, vocab]."""
+    def __call__(self, inputs, target_hidden, cache):
         h = self.embed_tokens(inputs)
         h_ctx = self.hidden_norm(self.fc(target_hidden))
-
         for layer, c in zip(self.layers, cache):
             h = layer(h, h_ctx, self.rope, c)
-
         return self.lm_head(self.norm(h))
+
+
+def load(model_id: str):
+    from mlx_lm import load as mlx_lm_load
+    return mlx_lm_load(model_id)
+
+
+def load_draft(draft_id: str) -> DFlashDraftModel:
+    path = Path(snapshot_download(draft_id, allow_patterns=["*.safetensors", "*.json"]))
+    cfg = json.loads((path / "config.json").read_text())
+    config = DFlashConfig(
+        hidden_size=cfg["hidden_size"],
+        num_hidden_layers=cfg["num_hidden_layers"],
+        num_attention_heads=cfg["num_attention_heads"],
+        num_key_value_heads=cfg["num_key_value_heads"],
+        head_dim=cfg["head_dim"],
+        intermediate_size=cfg["intermediate_size"],
+        vocab_size=cfg["vocab_size"],
+        rms_norm_eps=cfg["rms_norm_eps"],
+        rope_theta=cfg["rope_theta"],
+        max_position_embeddings=cfg["max_position_embeddings"],
+        block_size=cfg["block_size"],
+        target_layer_ids=tuple(cfg["dflash_config"]["target_layer_ids"]),
+        num_target_layers=cfg["num_target_layers"],
+        mask_token_id=cfg["dflash_config"]["mask_token_id"],
+    )
+    weights = {k: v for f in path.glob("*.safetensors") for k, v in mx.load(str(f)).items()}
+    model = DFlashDraftModel(config)
+    model.load_weights(list(weights.items()))
+    return model
+
+
+class _LayerHook:
+    def __init__(self, layer, idx, storage):
+        self._layer, self._idx, self._storage = layer, idx, storage
+
+    def __call__(self, *args, **kwargs):
+        self._storage[self._idx] = out = self._layer(*args, **kwargs)
+        return out
+
+    def __getattr__(self, name):
+        return getattr(self._layer, name)
+
+
+def _get_layers(model):
+    if hasattr(model, "model") and hasattr(model.model, "layers"):
+        return model.model.layers
+    if hasattr(model, "language_model") and hasattr(model.language_model, "layers"):
+        return model.language_model.layers
+    if hasattr(model, "layers"):
+        return model.layers
+    raise AttributeError(f"Cannot find layers in {type(model).__name__}")
+
+
+def _patch_model(model, layer_ids):
+    if hasattr(model, "_hidden_states"):
+        return
+    model._hidden_states = [None] * len(layer_ids)
+    layers = _get_layers(model)
+    for i, lid in enumerate(layer_ids):
+        layers[lid] = _LayerHook(layers[lid], i, model._hidden_states)
+
+
+class _GDNStateCapture:
+    def __init__(self):
+        self.conv_data = []
+        self._gdn_inputs = []
+        self._patch()
+
+    def _patch(self):
+        from mlx_lm.models.qwen3_5 import GatedDeltaNet
+        capture = self
+
+        def _capturing_gdn_call(self_layer, inputs, mask=None, cache=None):
+            B, S, _ = inputs.shape
+            if self_layer.sharding_group is not None:
+                from mlx_lm.models.qwen3_5 import sum_gradients
+                inputs = sum_gradients(self_layer.sharding_group)(inputs)
+            qkv = self_layer.in_proj_qkv(inputs)
+            z = self_layer.in_proj_z(inputs).reshape(B, S, self_layer.num_v_heads, self_layer.head_v_dim)
+            b, a = self_layer.in_proj_b(inputs), self_layer.in_proj_a(inputs)
+            conv_state = cache[0] if (cache is not None and cache[0] is not None) else mx.zeros((B, self_layer.conv_kernel_size - 1, self_layer.conv_dim), dtype=inputs.dtype)
+            if mask is not None:
+                qkv = mx.where(mask[..., None], qkv, 0)
+            conv_input = mx.concatenate([conv_state, qkv], axis=1)
+            capture.conv_data.append((conv_input, self_layer.conv_kernel_size))
+            if cache is not None:
+                cache[0] = conv_input[:, -(self_layer.conv_kernel_size - 1):]
+            conv_out = nn.silu(self_layer.conv1d(conv_input))
+            q, k, v = [
+                t.reshape(B, S, h, d)
+                for t, h, d in zip(
+                    mx.split(conv_out, [self_layer.key_dim, 2 * self_layer.key_dim], -1),
+                    [self_layer.num_k_heads, self_layer.num_k_heads, self_layer.num_v_heads],
+                    [self_layer.head_k_dim, self_layer.head_k_dim, self_layer.head_v_dim],
+                )
+            ]
+            state = cache[1] if cache else None
+            inv_scale = k.shape[-1] ** -0.5
+            q = (inv_scale ** 2) * mx.fast.rms_norm(q, None, 1e-6)
+            k = inv_scale * mx.fast.rms_norm(k, None, 1e-6)
+            capture._gdn_inputs.append((q, k, v, a, b, self_layer.A_log, self_layer.dt_bias, state, mask))
+            out, new_state = _gd_mod.gated_delta_update(
+                q, k, v, a, b, self_layer.A_log, self_layer.dt_bias, state, mask, use_kernel=True
+            )
+            if cache is not None:
+                cache[1] = new_state
+            out = self_layer.norm(out, z)
+            out = self_layer.out_proj(out.reshape(B, S, -1))
+            if self_layer.sharding_group is not None:
+                out = mx.distributed.all_sum(out, group=self_layer.sharding_group)
+            return out
+
+        GatedDeltaNet.__call__ = _capturing_gdn_call
+
+    def clear(self):
+        self.conv_data.clear()
+        self._gdn_inputs.clear()
+
+    def rollback(self, cache, accepted, trim):
+        j = 0
+        for c in cache:
+            if c.is_trimmable():
+                c.trim(trim)
+            else:
+                q, k, v, a, b, A_log, dt_bias, init_state, mask = self._gdn_inputs[j]
+                n = accepted + 1
+                _, state = _gd_mod.gated_delta_update(
+                    q[:, :n], k[:, :n], v[:, :n], a[:, :n], b[:, :n],
+                    A_log, dt_bias, init_state,
+                    None if mask is None else mask[:, :n],
+                    use_kernel=True,
+                )
+                c.cache[1] = state
+                conv_input, K = self.conv_data[j]
+                c.cache[0] = conv_input[:, accepted + 1 : accepted + K]
+                j += 1
+
+
+@dataclass
+class GenerationResponse:
+    text: str
+    tokens: List[int]
+    accepted: int
+    prompt_tokens: int
+    prompt_tps: float
+    generation_tokens: int
+    generation_tps: float
+    peak_memory: float
+    finish_reason: Optional[str] = None
+
+
+def _make_response(text, tokens, accepted, prompt_size, prompt_tps, n, tic, finish_reason=None):
+    return GenerationResponse(
+        text, tokens, accepted, prompt_size, prompt_tps,
+        n, n / (time.perf_counter() - tic), mx.get_peak_memory() / 1e9, finish_reason,
+    )
+
+
+def stream_generate(
+    model, draft, tokenizer, prompt,
+    block_size=None, max_tokens=256, temperature=0.0, sampler=None,
+):
+    _patch_model(model, draft.config.target_layer_ids)
+    block_size = block_size if block_size is not None else int(draft.config.block_size)
+    sampler = sampler or make_sampler(temp=temperature)
+
+    if not isinstance(tokenizer, TokenizerWrapper):
+        tokenizer = TokenizerWrapper(tokenizer)
+
+    if not isinstance(prompt, mx.array):
+        if isinstance(prompt, str):
+            add_special_tokens = tokenizer.bos_token is None or not prompt.startswith(tokenizer.bos_token)
+            prompt = tokenizer.encode(prompt, add_special_tokens=add_special_tokens)
+        prompt = mx.array(prompt)
+
+    detokenizer = tokenizer.detokenizer
+    mask_id = int(draft.config.mask_token_id)
+    tokens = prompt.tolist()
+
+    target_cache = make_prompt_cache(model)
+    draft_cache = make_prompt_cache(draft)
+    draft.bind(model)
+    _target_can_trim = can_trim_prompt_cache(target_cache)
+    _capture = _GDNStateCapture() if (not _target_can_trim and _HAS_GDN) else None
+
+    tic = time.perf_counter()
+    with mx.stream(generation_stream):
+        logits = model(prompt[None], target_cache)
+        hidden = mx.concatenate(model._hidden_states, axis=-1)
+    mx.eval(logits, hidden)
+    prompt_tps = prompt.size / (time.perf_counter() - tic)
+
+    tic = time.perf_counter()
+    token = sampler(logits[:, -1:])[0, 0].item()
+    tokens.append(token)
+    n = 1
+
+    if token in tokenizer.eos_token_ids:
+        detokenizer.add_token(token)
+        detokenizer.finalize()
+        yield _make_response(detokenizer.last_segment, [token], 1, prompt.size, prompt_tps, n, tic, "stop")
+        return
+
+    detokenizer.add_token(token)
+    yield _make_response(detokenizer.last_segment, [token], 1, prompt.size, prompt_tps, n, tic)
+
+    while n < max_tokens:
+        bs = min(block_size, max_tokens - n + 1)
+        if bs <= 1:
+            break
+
+        with mx.stream(generation_stream):
+            block = mx.array([[tokens[-1]] + [mask_id] * (bs - 1)])
+            draft_logits = draft(block, hidden, draft_cache)
+            if (trim_n := draft_cache[0].offset - (prompt.size + n - 1)) > 0:
+                trim_prompt_cache(draft_cache, trim_n)
+            draft_tokens = sampler(draft_logits[:, 1 - bs:])
+        mx.async_eval(draft_tokens)
+
+        if _capture is not None:
+            _capture.clear()
+        with mx.stream(generation_stream):
+            verify_input = mx.concatenate([mx.array([[tokens[-1]]]), draft_tokens], axis=1)
+            logits = model(verify_input, target_cache)
+            hidden = mx.concatenate(model._hidden_states, axis=-1)
+            target_tokens = sampler(logits)
+        mx.async_eval(target_tokens, hidden)
+
+        d_list, t_list = draft_tokens[0].tolist(), target_tokens[0].tolist()
+        accepted = next((i for i in range(len(d_list)) if d_list[i] != t_list[i]), len(d_list))
+        new_tokens = d_list[:accepted] + [t_list[accepted]]
+        new_tokens = new_tokens[:max_tokens - n]
+
+        eos_idx = next((i for i, t in enumerate(new_tokens) if t in tokenizer.eos_token_ids), None)
+        if eos_idx is not None:
+            new_tokens = new_tokens[:eos_idx + 1]
+            for t in new_tokens:
+                detokenizer.add_token(t)
+            detokenizer.finalize()
+            tokens.extend(new_tokens)
+            n += len(new_tokens)
+            yield _make_response(detokenizer.last_segment, new_tokens, accepted + 1, prompt.size, prompt_tps, n, tic, "stop")
+            return
+
+        for t in new_tokens:
+            detokenizer.add_token(t)
+        tokens.extend(new_tokens)
+        n += len(new_tokens)
+
+        if n % 256 == 0:
+            mx.clear_cache()
+
+        yield _make_response(detokenizer.last_segment, new_tokens, accepted + 1, prompt.size, prompt_tps, n, tic)
+
+        trim = bs - accepted - 1
+        if trim > 0:
+            if _target_can_trim:
+                trim_prompt_cache(target_cache, trim)
+            elif _capture is not None:
+                _capture.rollback(target_cache, accepted, trim)
+        hidden = hidden[:, :accepted + 1, :]
+
+    detokenizer.finalize()
+    yield _make_response(detokenizer.last_segment, [], 0, prompt.size, prompt_tps, n, tic, "length")
